created new sample code file
separated sample code to a separate file for modularity
